catch previously uncaught exception
Also for readHVRelays the declaration and definition's signature didn't
match. Fixed that too

+++ Source/Objects/Custom Hardware/SNO+/XL3/ORXL3Model.h
@@ -441,7 +441,7 @@ enum {
 
 - (void) setHVRelays:(unsigned long long)relayMask error:(unsigned long*)aError;
 - (void) setHVRelays:(unsigned long long)relayMask;
-- (void) readHVRelays:(unsigned long long*) relayMask isKnown:(BOOL*)isKnown;
+- (void) readHVRelays:(uint64_t*) relayMask isKnown:(BOOL*)isKnown;
 - (void) closeHVRelays;
 - (void) openHVRelays;
